fix: integration test
